requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
openapi-spec-validator 
pyyaml
dotenv
fastapi 
uvicorn
langgraph-checkpoint-postgres
psycopg[binary,pool]